farmercrophealthbackend/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
Pillow==10.2.0
Flask[async]==3.0.2
Flask-Cors==4.0.0
google-generativeai==0.7.1
python-dotenv==1.0.1
